feat(flags): integrate Vercel Flags Explorer discovery endpoint
Adopt the official `flags` SDK for the /.well-known/vercel/flags endpoint
so the Vercel Toolbar's Feature Flags panel can discover, describe, and
override every flag the app exposes.

- Replace the hand-rolled Bearer check with `verifyAccess` from `flags`,
  which validates a JWE access proof instead of a plaintext secret.
- Emit the `ApiData` shape the Toolbar expects: per-flag `options`
  (with labels), `origin`, `description`, `defaultValue`, and
  `declaredInCode: true`. Descriptions now come from the registry's
  human-readable `label` + `description` + release status.
- Split `featureRegistry.jsx` into a pure-data `featureRegistryData.js`
  (no JSX, no React, no icons) plus a thin view wrapper that reattaches
  lucide-react icons. The serverless function imports the data module
  directly, giving the Toolbar the full registry metadata without
  pulling React into the function bundle.
- Add `tests/unit/flagsExplorerEndpoint.test.js` to cover the auth gate
  and response shape, and guard the shared test setup so node-env tests
  can opt out of jsdom-only polyfills.

FLAGS_SECRET must be set in Vercel project env vars for the endpoint to
respond; the Toolbar panel handles provisioning the matching access
proof.
