+++ package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "@openfeature/react-sdk": "^1.2.1",
     "@tweakch/collection-grimm-top5": "*",
+    "flags": "^4.0.6",
     "i18next": "^26.0.4",
     "lucide-react": "1.7.0",
     "react": "18.2.0",
